Rename ExtensionPointFunctionComponent to ComponentLoader, add SiteEditorWrapper to add missing functionality from ExtensionPointClassComponent, and update types
